refactor/face-detection-folder: use index file for package-structure like
