+++ src/utils/faceDetection/index.ts
@@ -1,0 +1,2 @@
+export * from './plugins';
+export * from './types';
